refactor(browser): extract headless scheduler
